Compute/requirements.txt
streamlit>=1.32.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
anthropic>=0.40.0
python-dotenv>=1.0.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=1.12.0
lxml>=5.0.0
apscheduler>=3.10.0
python-docx>=1.1.0
plotly>=5.20.0
pandas>=2.0.0
